# swagger-collection.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Carsharing API",
    "description": "Simple carsharing api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/api/"
    }
  ],
...
  "paths": {
...
    "/bind-car": {
      "post": {
        "tags": [
          "carsharing api"
        ],
        "summary": "bind car to user",
        "description": "Bind car to user by car_id and user_id",
        "operationId": "bindCar",
        "requestBody": {
          "$ref": "#/components/requestBodies/BindCar"
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "The car is busy or the user has already taken the car"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/unbind-car": {
      "post": {
        "tags": [
          "carsharing api"
        ],
        "summary": "unbind car from user",
        "description": "Unbind car from user by user_id",
        "operationId": "unbindCar",
        "requestBody": {
          "$ref": "#/components/requestBodies/UnbindCar"
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "The user did not take the car"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    }
  },
  "components": {
...
    "requestBodies": {
...
      "BindCar": {
        "description": "Bind car to user",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "integer",
                  "example": 1
                },
                "car_id": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          }
        },
        "required": true
      },
      "UnbindCar": {
        "description": "Unbind car from user",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          }
        },
        "required": true
      }
    }
  }
}